+++ src/standalone/cline-core.ts
@@ -1,35 +1,29 @@
-import { activate } from "@/extension"
-import { Controller } from "@core/controller"
-import { CacheService } from "@core/storage/CacheService"
+import { initialize } from "@/common"
+import { WebviewProvider } from "@/core/webview"
+import { AuthHandler } from "@/hosts/external/AuthHandler"
+import { HostProvider } from "@/hosts/host-provider"
+import { DiffViewProvider } from "@/integrations/editor/DiffViewProvider"
 import { ExternalDiffViewProvider } from "@hosts/external/ExternalDiffviewProvider"
 import { ExternalWebviewProvider } from "@hosts/external/ExternalWebviewProvider"
 import { ExternalHostBridgeClientManager } from "@hosts/external/host-bridge-client-manager"
-import { HostProvider } from "@/hosts/host-provider"
 import { WebviewProviderType } from "@shared/webview/types"
-import { v4 as uuidv4 } from "uuid"
+import { startProtobusService } from "./protobus-service"
 import { log } from "./utils"
 import { extensionContext } from "./vscode-context"
-import { startProtobusService } from "./protobus-service"
-import { AuthHandler } from "@/hosts/external/AuthHandler"
-import { WebviewProvider } from "@/core/webview"
-import { DiffViewProvider } from "@/integrations/editor/DiffViewProvider"
 
 async function main() {
 	log("\n\n\nStarting cline-core service...\n\n\n")
-
-	AuthHandler.getInstance().setEnabled(true)
 
 	setupHostProvider()
 
 	// Set up global error handlers to prevent process crashes
 	setupGlobalErrorHandlers()
 
-	activate(extensionContext)
-	// Create and initialize cache service
+	const webviewProvider = await initialize(extensionContext)
 
-	// Create controller with cache service
-	const controller = new Controller(extensionContext, uuidv4())
-	startProtobusService(controller)
+	AuthHandler.getInstance().setEnabled(true)
+
+	startProtobusService(webviewProvider.controller)
 }
 
 function setupHostProvider() {
